[ui]: reorganising theme, change and standardised fonts
