+++ public/index.html
@@ -16,7 +16,7 @@
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link
-      href="https://fonts.googleapis.com/css?family=Merriweather|Quicksand&display=swap"
+      href="https://fonts.googleapis.com/css?family=Montserrat|Varela+Round&display=swap"
       rel="stylesheet"
     />
     <!--
