+++ src/components/MoneyTypography/index.ts
@@ -1,0 +1,2 @@
+import MoneyTypography from "./MoneyTypography";
+export default MoneyTypography;
