+++ src/components/MoneyTypography/styles.ts
@@ -1,0 +1,11 @@
+import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
+
+const useStyles = makeStyles<Theme>(theme =>
+  createStyles({
+    root: {
+      fontFamily: "Varela Round, Montserrat"
+    }
+  })
+);
+
+export default useStyles;
