+++ src/components/MoneyTypography/types.d.ts
@@ -1,0 +1,7 @@
+import { ThemeStyle } from "@material-ui/core/styles/createTypography";
+
+export default interface MoneyTypographyProps {
+  value?: number;
+
+  variant?: ThemeStyle;
+}
